chore: move RenderEpisodesButton to atoms
